+++ src/engine/components/quick_bar.cpp
@@ -5,43 +5,68 @@
 
 #include "../app.hpp"
 #include "../result.hpp"
-#include "component.hpp"
+#include "button.hpp"
 #include "sprite.hpp"
+#include "ui_component.hpp"
 
 #include <raylib.h>
 
 #include <algorithm>
+#include <cassert>
 #include <cstddef>
 #include <memory>
 #include <string>
+#include <format>
 
 namespace engine {
 
-auto quick_bar::init(app &app, const std::string &sprite_sheet, float gap) -> result<> {
-	if(const auto err = component::init(app).unwrap(); err) {
-		return error("failed to initialize base component", *err);
+result<> quick_bar::init(app &app) {
+	return ui_component::init(app);
+}
+
+auto quick_bar::init(app &app,
+					 const std::string &sprite_sheet,
+					 const Color normal_color,
+					 const Color hover_color,
+					 const float gap) -> result<> {
+	if(const auto err = init(app).unwrap(); err) {
+		return error("failed to initialize base quick_bar component", *err);
 	}
 
 	sprite_sheet_ = sprite_sheet;
 	gap_ = gap;
+	normal_color_ = normal_color;
+	hover_color_ = hover_color;
 
 	return true;
 }
 
-auto quick_bar::add_sprite(const std::string &frame_name) -> result<size_t> {
+auto quick_bar::add_button(const std::string &frame_name) -> result<size_t> {
 	auto sprite_ptr = std::make_shared<sprite>();
 	if(const auto err = sprite_ptr->init(get_app(), sprite_sheet_, frame_name).unwrap(); err) {
 		return error("failed to initialize sprite in quick_bar", *err);
 	}
-	sprite_ptr->set_tint(Color(255, 255, 255, 60));
+	assert(sprite_ptr->get_pivot().x == 0.5F && sprite_ptr->get_pivot().y == 0.5F && "sprite pivot must be centered");
+
+	sprite_ptr->set_tint(normal_color_);
 	sprites_.emplace_back(sprite_ptr);
 	recalculate();
 
 	return sprite_ptr->get_id();
 }
 
+auto quick_bar::set_button_frame_name(size_t button, const std::string &frame_name) const -> result<> {
+	for(auto &sprite_ptr: sprites_) {
+		if(sprite_ptr->get_id() == button) {
+			sprite_ptr->set_frame_name(frame_name);
+			return true;
+		}
+	}
+	return error(std::format("can not set button frame, button id not found in quick_bar: {}", button));
+}
+
 auto quick_bar::set_position(const Vector2 &pos) -> void {
-	component::set_position(pos);
+	ui_component::set_position(pos);
 	recalculate();
 }
 
@@ -53,7 +78,7 @@ auto quick_bar::end() -> result<> {
 	}
 	sprites_.clear();
 
-	return component::end();
+	return ui_component::end();
 }
 
 auto quick_bar::update(float delta) -> result<> {
@@ -63,7 +88,23 @@ auto quick_bar::update(float delta) -> result<> {
 		}
 	}
 
-	return component::update(delta);
+	const auto mouse = GetMousePosition();
+	for(const auto &sprite_ptr: sprites_) {
+		if(sprite_ptr->point_inside(mouse)) {
+			sprite_ptr->set_tint(hover_color_);
+			if(IsMouseButtonPressed(MOUSE_LEFT_BUTTON)) {
+				if(const auto err = play_click_sound().unwrap(); err) {
+					return error("failed to play click sound", *err);
+				}
+				get_app().post_event(button::click{.id = sprite_ptr->get_id()});
+			}
+
+		} else {
+			sprite_ptr->set_tint(normal_color_);
+		}
+	}
+
+	return ui_component::update(delta);
 }
 
 auto quick_bar::draw() -> result<> {
@@ -73,7 +114,7 @@ auto quick_bar::draw() -> result<> {
 		}
 	}
 
-	return component::draw();
+	return ui_component::draw();
 }
 
 auto quick_bar::recalculate() -> void {